// apps/user-service/src/modules/register/dtos/restaurant.dto.ts
export type IRestaurantRegisterDTO = {
  username: string;
  fullName: string;
  email: string;
  password: string;
  profileImageURL: string;
  street: string;
  number: number;
  neighborhood: string;
  city: string;
  country: string;
  zipCode: string;
  state: string;
  latitude: number;
  longitude: number;
};